Adapted test infrastructure based on recent JADNC refactorings
